+++ client/components/registration/registrationForm.spec.js
@@ -2,27 +2,28 @@ import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 
-import LoginForm from './LoginForm.jsx';
+import RegistrationForm from './RegistrationForm.jsx';
 import TextField from '../common/TextField.jsx';
 import RaisedButton from 'material-ui/RaisedButton';
 
-describe('<LoginForm />', () => {
+describe('<RegistrationForm />', () => {
   const props = {
+    errors: {},
     onChange() {},
     login() {},
   };
 
-  it('should display two input fields', () => {
-    const wrapper = shallow(<LoginForm {...props} />);
-    const expected = 2;
+  it('should display seven input fields', () => {
+    const wrapper = shallow(<RegistrationForm {...props} />);
+    const expected = 7;
 
     const actual = wrapper.find(TextField).length;
     expect(actual).equals(expected);
   });
 
-  it('should display a button that states "Log In"', () => {
-    const wrapper = shallow(<LoginForm {...props} />);
-    const expected = 'Log In';
+  it('should display a button that states "Register"', () => {
+    const wrapper = shallow(<RegistrationForm {...props} />);
+    const expected = 'Register';
 
     const actual = wrapper.find(RaisedButton).props().label;
     expect(actual).equals(expected);
@@ -34,47 +35,55 @@ describe('<LoginForm />', () => {
       hidden: true,
     };
 
-    const wrapper = shallow(<LoginForm {...updatedProps} />);
+    const wrapper = shallow(<RegistrationForm {...updatedProps} />);
     const expected = 'none';
 
     const actual = wrapper.first().props().style.display;
+
     expect(actual).equals(expected);
   });
 
-  it('should set the display to initial on the form if the hidden prop is passed in as true', () => {
+  it('should set the display to initial on the form if the server error prop is passed in as true', () => {
     const updatedProps = {
       ...props,
-      hidden: false,
+      errors: {
+        server: true,
+      },
     };
 
-    const wrapper = shallow(<LoginForm {...updatedProps} />);
+    const wrapper = shallow(<RegistrationForm {...updatedProps} />);
     const expected = 'initial';
 
     const actual = wrapper.first().props().style.display;
+
     expect(actual).equals(expected);
   });
 
-  it('should display the error text if the errors prop is passed in as true', () => {
+  it('should not display the error text if the server errors state is false', () => {
     const updatedProps = {
       ...props,
-      errors: true,
+      errors: {
+        server: false,
+      },
     };
 
-    const wrapper = shallow(<LoginForm {...updatedProps} />);
-    const expected = 'initial';
+    const wrapper = shallow(<RegistrationForm {...updatedProps} />);
+    const expected = 'none';
 
     const actual = wrapper.find('.error-text').props().style.display;
     expect(actual).equals(expected);
   });
 
-  it('should display the error text of "Invalid email address or password" for generic errors', () => {
+  it('should display the text, "Cannot register, please try again later" if the error text is visible', () => {
     const updatedProps = {
       ...props,
-      errors: true,
+      errors: {
+        server: true,
+      },
     };
 
-    const wrapper = shallow(<LoginForm {...updatedProps} />);
-    const expected = 'Invalid email address or password';
+    const expected = 'Cannot register, please try again later';
+    const wrapper = shallow(<RegistrationForm {...updatedProps} />);
 
     const actual = wrapper.find('.error-text').text();
     expect(actual).equals(expected);