+++ client/utilities/endpoints.js
@@ -3,6 +3,7 @@ const protocol = window.location.protocol;
 const baseEndpoint = `${protocol}//${hostLocation}`;
 
 export const ACCOUNT_ENDPOINT = `${baseEndpoint}/api/v1/account`;
+export const REGISTRATION_ENDPOINT = `${baseEndpoint}/api/v1/registration`;
 export const VERSION_ENDPOINT = `${baseEndpoint}/api/v1/version`;
 export const IMAGE_ENDPOINT = `${baseEndpoint}/api/v1/images`;
 export const ITEM_ENDPOINT = `${baseEndpoint}/api/v1/items`;